refactor: move privilege-checking code
