+++ internal/config/config_read.go
@@ -8,9 +8,11 @@ import (
 	"github.com/favonia/cloudflare-ddns/internal/provider"
 )
 
-// ReadEnv calls the relevant readers to read all relevant environment variables except TZ
-// and update relevant fields. One should subsequently call [Config.NormalizeConfig]
-// to maintain invariants across different fields.
+// ReadEnv calls the relevant readers to read all relevant environment variables except
+// - timezone (TZ)
+// - privileges-related ones (PGID and PUID)
+// - output-related ones (QUIET and EMOJI)
+// One should subsequently call [Config.NormalizeConfig] to restore invariants across fields.
 func (c *Config) ReadEnv(ppfmt pp.PP) bool {
 	if ppfmt.IsEnabledFor(pp.Info) {
 		ppfmt.Infof(pp.EmojiEnvVars, "Reading settings . . .")
@@ -34,9 +36,6 @@ func (c *Config) ReadEnv(ppfmt pp.PP) bool {
 		!ReadAndAppendShoutrrrURL(ppfmt, "SHOUTRRR", &c.Notifiers) {
 		return false
 	}
-
-	CheckIgnoredLinuxID(ppfmt, "PUID", "user")
-	CheckIgnoredLinuxID(ppfmt, "PGID", "group")
 
 	return true
 }
