+++ internal/config/env_base_test.go
@@ -37,6 +37,13 @@ func unset(t *testing.T, keys ...string) {
 	}
 }
 
+func urlMustParse(t *testing.T, u string) *url.URL {
+	t.Helper()
+	url, err := url.Parse(u)
+	require.NoError(t, err)
+	return url
+}
+
 //nolint:paralleltest // environment vars are global
 func TestGetenv(t *testing.T) {
 	key := keyPrefix + "VAR"
@@ -122,98 +129,6 @@ func TestReadString(t *testing.T) {
 			ok := config.ReadString(mockPP, key, &field)
 			require.Equal(t, tc.ok, ok)
 			require.Equal(t, tc.newField, field)
-		})
-	}
-}
-
-//nolint:paralleltest // environment vars are global
-func TestReadEmoji(t *testing.T) {
-	key := keyPrefix + "EMOJI"
-	for name, tc := range map[string]struct {
-		set           bool
-		val           string
-		ok            bool
-		prepareMockPP func(*mocks.MockPP)
-	}{
-		"nil":   {false, "", true, nil},
-		"empty": {true, " ", true, nil},
-		"true": {
-			true, " true", true,
-			func(m *mocks.MockPP) {
-				m.EXPECT().SetEmoji(true)
-			},
-		},
-		"false": {
-			true, "    false ", true,
-			func(m *mocks.MockPP) {
-				m.EXPECT().SetEmoji(false)
-			},
-		},
-		"illform": {
-			true, "weird", false,
-			func(m *mocks.MockPP) {
-				m.EXPECT().Errorf(pp.EmojiUserError, "%s (%q) is not a boolean: %v", key, "weird", gomock.Any())
-			},
-		},
-	} {
-		t.Run(name, func(t *testing.T) {
-			set(t, key, tc.set, tc.val)
-			mockCtrl := gomock.NewController(t)
-			mockPP := mocks.NewMockPP(mockCtrl)
-			if tc.prepareMockPP != nil {
-				tc.prepareMockPP(mockPP)
-			}
-
-			var wrappedPP pp.PP = mockPP
-
-			ok := config.ReadEmoji(key, &wrappedPP)
-			require.Equal(t, tc.ok, ok)
-		})
-	}
-}
-
-//nolint:paralleltest // environment vars are global
-func TestReadQuiet(t *testing.T) {
-	key := keyPrefix + "QUIET"
-	for name, tc := range map[string]struct {
-		set           bool
-		val           string
-		ok            bool
-		prepareMockPP func(*mocks.MockPP)
-	}{
-		"nil":   {false, "", true, nil},
-		"empty": {true, " ", true, nil},
-		"true": {
-			true, " true", true,
-			func(m *mocks.MockPP) {
-				m.EXPECT().SetVerbosity(pp.Notice)
-			},
-		},
-		"false": {
-			true, "    false ", true,
-			func(m *mocks.MockPP) {
-				m.EXPECT().SetVerbosity(pp.Info)
-			},
-		},
-		"illform": {
-			true, "weird", false,
-			func(m *mocks.MockPP) {
-				m.EXPECT().Errorf(pp.EmojiUserError, "%s (%q) is not a boolean: %v", key, "weird", gomock.Any())
-			},
-		},
-	} {
-		t.Run(name, func(t *testing.T) {
-			set(t, key, tc.set, tc.val)
-			mockCtrl := gomock.NewController(t)
-			mockPP := mocks.NewMockPP(mockCtrl)
-			if tc.prepareMockPP != nil {
-				tc.prepareMockPP(mockPP)
-			}
-
-			var wrappedPP pp.PP = mockPP
-
-			ok := config.ReadQuiet(key, &wrappedPP)
-			require.Equal(t, tc.ok, ok)
 		})
 	}
 }
@@ -530,11 +445,4 @@ func TestReadCron(t *testing.T) {
 			require.Equal(t, tc.newField, field)
 		})
 	}
-}
-
-func urlMustParse(t *testing.T, u string) *url.URL {
-	t.Helper()
-	url, err := url.Parse(u)
-	require.NoError(t, err)
-	return url
 }
